Disable build of GObject Introspection, fix cross compile
Not trivial in a cross setting as BUILD and HOST dev packages
must be installed side by side and the upstream debian ones
doesn't support it.

Confirmed that upstream python-gst-1.0, that depends on
gir1.2-gstreamer-1.0, can be installed and used alongside
our binary debs so we're not losing language bindings.

Change-Id: Ia77d8bfd5c201d074faea40ad657e8dbebdc6e12